feat: 반정규화 후 Product 테이블 조회 쿼리 테스트
- ddl-auto:none 후 db 스키마 변경 및 상품좋아요 백필 수행

```sql
ALTER TABLE product ADD COLUMN like_count BIGINT NOT NULL DEFAULT 0;

UPDATE product p
JOIN (
  SELECT product_id, COUNT(*) AS cnt
  FROM product_like
  GROUP BY product_id
) pl ON pl.product_id = p.id
SET p.like_count = pl.cnt;
```

+++ apps/commerce-api/src/main/java/com/loopers/domain/product/ProductModel.java
@@ -36,7 +36,7 @@ public class ProductModel extends BaseEntity {
     private BrandModel brand;
 
     @Column(name = "like_count")
-    private int likeCount;
+    private long likeCount;
 
     protected ProductModel() {}
